% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommentManager.R
\name{CommentManager}
\alias{CommentManager}
\title{R markdown manager for automating annotated commenting on text}
\description{
An R6 class for the management of comments. Includes option for checking
  for pandoc compilation to Word docx files to use the Word feature
  of comments in margins.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CommentManager-new}{\code{CommentManager$new()}}
\item \href{#method-CommentManager-comment}{\code{CommentManager$comment()}}
\item \href{#method-CommentManager-clone}{\code{CommentManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-CommentManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentManager$new(
  commentCount = 0,
  author = "Anonymous",
  timeDefault = "1970-01-01T00:00:00Z",
  checkWord = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentManager-comment"></a>}}
\if{latex}{\out{\hypertarget{method-CommentManager-comment}{}}}
\subsection{Method \code{comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentManager$comment(
  text,
  comment,
  author = self$author,
  time = NULL,
  id = NULL,
  checkWord = self$checkWord
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CommentManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CommentManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommentManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
