% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquationNumberManager.R
\name{EquationNumberManager}
\alias{EquationNumberManager}
\title{R markdown manager for automating equation numbers}
\description{
An R6 class for the management of automated equation numbers,
  including a feature for cross-referencing previously defined equations
  by number.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EquationNumberManager-new}{\code{EquationNumberManager$new()}}
\item \href{#method-EquationNumberManager-addEq}{\code{EquationNumberManager$addEq()}}
\item \href{#method-EquationNumberManager-getNum}{\code{EquationNumberManager$getNum()}}
\item \href{#method-EquationNumberManager-clone}{\code{EquationNumberManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EquationNumberManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-EquationNumberManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EquationNumberManager$new(counter = 0, parens = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EquationNumberManager-addEq"></a>}}
\if{latex}{\out{\hypertarget{method-EquationNumberManager-addEq}{}}}
\subsection{Method \code{addEq()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EquationNumberManager$addEq(name = NULL, counter = NULL, parens = self$parens)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EquationNumberManager-getNum"></a>}}
\if{latex}{\out{\hypertarget{method-EquationNumberManager-getNum}{}}}
\subsection{Method \code{getNum()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EquationNumberManager$getNum(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EquationNumberManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EquationNumberManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EquationNumberManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
