% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigureManager.R
\name{FigureManager}
\alias{FigureManager}
\title{R markdown manager for automating figure references}
\description{
An R6 class for the management of figures, including: figure numbering,
  cross-references of figure numbers, and exporting figures with captions
  to R markdown.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{figCount}}{An integer counter for the total count of citable figures}

\item{\code{figList}}{A list where each element is information about a citable figure}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FigureManager-new}{\code{FigureManager$new()}}
\item \href{#method-FigureManager-addFig}{\code{FigureManager$addFig()}}
\item \href{#method-FigureManager-citeFig}{\code{FigureManager$citeFig()}}
\item \href{#method-FigureManager-catMarkdownAll}{\code{FigureManager$catMarkdownAll()}}
\item \href{#method-FigureManager-catCaptionsAll}{\code{FigureManager$catCaptionsAll()}}
\item \href{#method-FigureManager-clone}{\code{FigureManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FigureManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-FigureManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FigureManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FigureManager-addFig"></a>}}
\if{latex}{\out{\hypertarget{method-FigureManager-addFig}{}}}
\subsection{Method \code{addFig()}}{
Adds the information for a figure to the citable list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FigureManager$addFig(name, caption, path, fun = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string with the unique name of the figure}

\item{\code{caption}}{Character string with the figure caption.
Markdown formatting is allowed.}

\item{\code{path}}{Character string with the path to the image file for the figure.}

\item{\code{fun}}{Optional function for generating the markdown for rendering the
figure and its caption.
Default value is NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the index of the figure in the list of
  citable figures
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FigureManager-citeFig"></a>}}
\if{latex}{\out{\hypertarget{method-FigureManager-citeFig}{}}}
\subsection{Method \code{citeFig()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FigureManager$citeFig(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FigureManager-catMarkdownAll"></a>}}
\if{latex}{\out{\hypertarget{method-FigureManager-catMarkdownAll}{}}}
\subsection{Method \code{catMarkdownAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FigureManager$catMarkdownAll(
  newpage = TRUE,
  as.string = FALSE,
  type = c("output", "message"),
  split = FALSE,
  md.heading = "##"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FigureManager-catCaptionsAll"></a>}}
\if{latex}{\out{\hypertarget{method-FigureManager-catCaptionsAll}{}}}
\subsection{Method \code{catCaptionsAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FigureManager$catCaptionsAll(
  as.string = FALSE,
  type = c("output", "message"),
  split = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FigureManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FigureManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FigureManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
