% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SectionNumberManager.R
\name{SectionNumberManager}
\alias{SectionNumberManager}
\title{R markdown manager for automating hierarchical section numbers}
\description{
An R6 class for the management of automated hierarchical section numbers.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SectionNumberManager-new}{\code{SectionNumberManager$new()}}
\item \href{#method-SectionNumberManager-addSection}{\code{SectionNumberManager$addSection()}}
\item \href{#method-SectionNumberManager-getSectionNumber}{\code{SectionNumberManager$getSectionNumber()}}
\item \href{#method-SectionNumberManager-clone}{\code{SectionNumberManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectionNumberManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-SectionNumberManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionNumberManager$new(counters = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectionNumberManager-addSection"></a>}}
\if{latex}{\out{\hypertarget{method-SectionNumberManager-addSection}{}}}
\subsection{Method \code{addSection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionNumberManager$addSection(level, reset = 0, counters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectionNumberManager-getSectionNumber"></a>}}
\if{latex}{\out{\hypertarget{method-SectionNumberManager-getSectionNumber}{}}}
\subsection{Method \code{getSectionNumber()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionNumberManager$getSectionNumber()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SectionNumberManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SectionNumberManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SectionNumberManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
