% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableManager.R
\name{TableManager}
\alias{TableManager}
\title{R markdown manager for automating table references}
\description{
An R6 class for the management of tables, including: table numbering,
  cross-references of table numbers, and exporting tables with captions
  to R markdown.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{count}}{An integer counter for the total count of citable tables}

\item{\code{list}}{A list where each element is information about a citable table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableManager-new}{\code{TableManager$new()}}
\item \href{#method-TableManager-add}{\code{TableManager$add()}}
\item \href{#method-TableManager-cite}{\code{TableManager$cite()}}
\item \href{#method-TableManager-catMarkdownAll}{\code{TableManager$catMarkdownAll()}}
\item \href{#method-TableManager-catCaptionsAll}{\code{TableManager$catCaptionsAll()}}
\item \href{#method-TableManager-clone}{\code{TableManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableManager-add"></a>}}
\if{latex}{\out{\hypertarget{method-TableManager-add}{}}}
\subsection{Method \code{add()}}{
Adds the information for a table to the citable list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableManager$add(name, caption, df, fun = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string with the unique name of the table}

\item{\code{caption}}{Character string with the table caption.
Markdown formatting is allowed.}

\item{\code{df}}{Character string with the data frame with table data}

\item{\code{fun}}{Optional function for generating the markdown for rendering the
figure and its caption.
Default value is NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns the index of the table in the list of
  citable figures
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableManager-cite"></a>}}
\if{latex}{\out{\hypertarget{method-TableManager-cite}{}}}
\subsection{Method \code{cite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableManager$cite(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableManager-catMarkdownAll"></a>}}
\if{latex}{\out{\hypertarget{method-TableManager-catMarkdownAll}{}}}
\subsection{Method \code{catMarkdownAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableManager$catMarkdownAll(
  newpage = TRUE,
  as.string = FALSE,
  type = c("output", "message"),
  split = FALSE,
  md.heading = "##"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableManager-catCaptionsAll"></a>}}
\if{latex}{\out{\hypertarget{method-TableManager-catCaptionsAll}{}}}
\subsection{Method \code{catCaptionsAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableManager$catCaptionsAll(
  as.string = FALSE,
  type = c("output", "message"),
  split = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
